unit 6 end of week

+++ challenges/zookeeper-mongodb/controllers/animal.controller.js
@@ -1,15 +1,17 @@
 const router = require("express").Router();
 const Animal = require("../models/animal.model");
+const validateSession = require("../middleware/validate-session");
 
 
 // add new animal
-router.post("/create", async (req, res) => {
+router.post("/create",validateSession,  async (req, res) => {
     try {
         // preppering the animal object to be saved to the database 
         const animal = new Animal({
             name: req.body.name,
             legNumber: req.body.legNumber,
             predator: req.body.predator,
+            userId: req.user._id,
         })
         // we need to save the data
         const newAnimal = await animal.save();
@@ -24,7 +26,19 @@ router.post("/create", async (req, res) => {
     }
 })
 
-router.get("/", async (req, res) => {
+// get all animals by user
+router.get('/myanimals', validateSession, async(req, res) => {
+
+    try {
+      const animal = await Animal.find({userId: req.user._id})
+      res.json({ animal: animal, message: 'success' })
+    } catch(error) {
+      res.json({ message: error.message })
+    }
+})
+
+
+router.get("/", validateSession, async (req, res) => {
     try {
         const animal = await Animal.find();
         res.status(202).json({
@@ -37,7 +51,7 @@ router.get("/", async (req, res) => {
 })
 
 // get one animal
-router.get("/:id", async (req, res) => {
+router.get("/:id", validateSession, async (req, res) => {
     try {
         const animal = await Animal.findById({ _id: req.params.id });
         res.status(202).json({
@@ -50,7 +64,7 @@ router.get("/:id", async (req, res) => {
 })
 
 // Update
-router.patch("/update/:id", async (req, res) => {
+router.patch("/update/:id", validateSession, async (req, res) => {
     try {
         const filter = { _id: req.params.id }
 
@@ -59,7 +73,7 @@ router.patch("/update/:id", async (req, res) => {
 
         const animal = await Animal.findOneAndUpdate(filter, update, returnOptions);
         
-        res.json({message:"animal updated", animal:animal})
+        res.json({message: animal ?"animal updated" : "there is no animal by that id", animal:animal})
 
     }
      catch (error) {
@@ -68,9 +82,14 @@ router.patch("/update/:id", async (req, res) => {
 })
 
 // delete 
-router.delete("/delete/:id", async (req, res) => {
+router.delete("/delete/:id", validateSession, async (req, res) => {
     try {
         const animalToDelete = await Animal.findById({ _id: req.params.id });
+        if (!animalToDelete) throw new Error("Animal does not exist")
+        const isValidOwner = req.user._id == animalToDelete.userId
+        if (!isValidOwner) {
+            throw new Error("The id supplied for this animal is not owned by this user. Animal wasn't deleted")
+        }
         const deletedAnimal = await Animal.deleteOne({ _id: req.params.id });
         res.json({
         animalThatWasDeleted: animalToDelete,
